+++ src/app/core/modelo/options.model.ts
@@ -1,6 +1,0 @@
-import { HttpHeaders, HttpParams } from "@angular/common/http";
-
-export interface Options {
-  headers?: HttpHeaders;
-  params?: HttpParams;
-}