+++ src/app/feature/home/shared/services/favorite/favorite.service.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@angular/core';
+import { Character } from '@home/shared/model/character.model';
 
 @Injectable({
   providedIn: 'root'
@@ -15,7 +16,29 @@ export class FavoriteService {
     localStorage.setItem('favorites',JSON.stringify(data));
   }
 
-  randomDate(date1:string, date2:string){
+  add(data: Character): boolean {
+    let favorites = JSON.parse(localStorage.getItem('favorites')!);
+    let canAdd = favorites.filter((e: {id: number}) => e.id == data.id);
+    if(!canAdd) {
+      favorites = [...favorites, data];
+      localStorage.removeItem('favorites');
+      localStorage.setItem('favorites',JSON.stringify(favorites));
+      return true;
+    } else {
+      return false;
+    }
+  }
+
+  setRandom(response: any): void {
+    let favorites = JSON.parse(localStorage.getItem('favorites')!) || [];
+    for (let i = 0; i < response.data.results.length; i++) {
+      favorites = [...favorites,response.data.results[i]]
+    }
+    localStorage.removeItem('favorites');
+    localStorage.setItem('favorites',JSON.stringify(favorites));
+  }
+
+  randomDate(date1:string, date2:string): string{
     let newDate1 = (new Date(date1).getTime()).toString();
     let newDate2 = (new Date(date2).getTime()).toString();
     let date = new Date(this.getRandomDate(parseInt(newDate2), parseInt(newDate1)));
@@ -23,11 +46,11 @@ export class FavoriteService {
     return this.formatDate(date);
   }
 
-  getRandomDate(min:number, max:number) {
+  getRandomDate(min:number, max:number): number {
     return Math.random() * (max - min) + min;
   }
 
-  formatDate(date:any) {
+  formatDate(date:any): string {
     let d = new Date(date),
         month = '' + (d.getMonth() + 1),
         day = '' + d.getDate(),
@@ -41,7 +64,7 @@ export class FavoriteService {
     return [year, month, day].join('-');
   }
 
-  getStringDate() {
+  getStringDate(): string {
     let date1 = this.randomDate('03-08-2010', '03-10-2021');
     let date2 = this.randomDate('03-08-2000', '02-10-2010');
     return `${date2},${date1}`;